PyroGraph: Added symbol only comparison plotting function, fixed Smyth scaling factor, updated styles file to replace None for colors with w for white.
Former-commit-id: 56f8925bef656ca9edfb9605961bebd4a76a0d75 [formerly a1bf2956bf79e6d092a7435a1af8883844559530]
Former-commit-id: c0ede833056182ef9476964817b94a6c58959cef
Former-commit-id: 1df1865ef72788ce28419e540c400d3a77bbd860

+++ Utilities/PyroGraph/PyroGraph.py
@@ -159,7 +159,8 @@ if process_set == 1 or process_set == 3:
 # Clean up temp files
 for file_name in config_files:
     os.remove(file_name+'_object.pkl')
-os.remove('scatter_data_dict_object.pkl')
+if process_set == 1 or process_set == 3:
+    os.remove('scatter_data_dict_object.pkl')
 os.remove(config_file_name+'_object.pkl')
 
 ## Report Completion.
